removed unused validate file
